requirements.txt
web3==6.15.1
requests==2.31.0
python-dotenv==1.0.1
pandas==2.2.1
tqdm==4.66.2 